backend/requirements.txt
# Twitter Social Media Dashboard - Backend Dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
tweepy==4.14.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
